+++ packages/react/src/index.ts
@@ -1,14 +1,14 @@
-export { ActiveProps } from './Active';
-export { BlockProps } from './Block';
-export { CuriBaseProps } from './CuriBase';
-export { CuriousProps, CuriousComponent } from './curious';
-export { CuriProps, CuriContext } from './interface';
-export { LinkProps, LinkState, ActiveLink } from './Link';
+export { ActiveProps } from "./Active";
+export { BlockProps } from "./Block";
+export { CuriBaseProps } from "./CuriBase";
+export { CuriousProps, CuriousRenderProps } from "./Curious";
+export { CuriProps, CuriContext } from "./interface";
+export { LinkProps, LinkState, ActiveLink } from "./Link";
 
-import Active from './Active';
-import Block from './Block';
-import curious from './curious';
-import Link from './Link';
-import CuriBase from './CuriBase';
+import Active from "./Active";
+import Block from "./Block";
+import Curious from "./Curious";
+import Link from "./Link";
+import CuriBase from "./CuriBase";
 
-export { Active, Block, curious, Link, CuriBase };
+export { Active, Block, Curious, Link, CuriBase };