+++ packages/react/CHANGELOG.md
@@ -1,5 +1,6 @@
 ## Next
 
+* Switch from `curious` higher-order component to `<Curious>` with a render function.
 * Rename `<Link>`'s `curi` prop to `router`.
 
 ## 1.0.0-beta.15
@@ -11,7 +12,7 @@
 ## 1.0.0-beta.14
 
 * Make passing the `action` to `<CuriBase>` optional (it will default to `POP`).
-* Rename `<Navigator>` to `<CuriBase>`. 
+* Rename `<Navigator>` to `<CuriBase>`.
 * Add `extra` and `details` props to the `<Active>` component.
 * Add an `extra` function to the `<Link active>` object, which is a function that can add additional active checking using the current location and the details passed to the `<Link>`. This allows you to mark a link as active based on its `query` and/or `hash`.
 * Simpler `<Link anchor>` type.